requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# LLM integration
openai>=1.0.0
anthropic>=0.7.0  # Optional: for Claude API

# Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0  # Required by sentence-transformers
scikit-learn>=1.3.0  # For similarity calculations

# Network analysis
networkx>=3.0

# Data validation (optional but recommended)
pydantic>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0  # For environment variable management
tqdm>=4.65.0  # Progress bars

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

